+++ shared/src/main/scala/com/nawforce/pkgforce/diagnostics/IssueLogger.scala
@@ -108,7 +108,7 @@ class IssueLogger(val externalPathFilter: Option[PathLike => Boolean] = None)
     issuesForFilesInternal(internalPaths, includeWarnings, maxIssuesPerFile).toArray
   }
 
-  /** This is specialised accessor to aid PMD integration. */
+  /** This is a specialised accessor for issues that fall within a code range. */
   override def issuesForFileLocation(path: String, location: IssueLocation): Array[APIIssue] = {
     issuesForFileLocationInternal(Path(path), location)
   }